33504 - Refatoração inicial do questionário dinâmico (antigo encaminhamentoAEE)
