+++ src/SME.SGP.WebClient/src/redux/modulos/questionarioDinamico/actions.js
@@ -1,0 +1,9 @@
+export const setLimparDadosQuestionarioDinamico = payload => ({
+  type: '@questionarioDinamico/setLimparDadosQuestionarioDinamico',
+  payload,
+});
+
+export const setFormsQuestionarioDinamico = payload => ({
+  type: '@questionarioDinamico/setFormsQuestionarioDinamico',
+  payload,
+});
