+++ src/SME.SGP.WebClient/src/redux/modulos/reducers.js
@@ -29,6 +29,7 @@ import registroIndividual from './registroIndividual/reducers';
 import collapseLocalizarEstudante from './collapseLocalizarEstudante/reducers';
 import acompanhamentoFrequencia from './acompanhamentoFrequencia/reducers';
 import objectCardEstudante from './objectCardEstudante/reducers';
+import questionarioDinamico from './questionarioDinamico/reducers';
 
 const reducers = combineReducers({
   navegacao,
@@ -60,6 +61,7 @@ const reducers = combineReducers({
   collapseLocalizarEstudante,
   acompanhamentoFrequencia,
   objectCardEstudante,
+  questionarioDinamico,
 });
 
 const rootReducer = (state, action) => {
